refactor: best-soeat-card 폴더 제거 및 SosoTalkCardCompact 이동
 - 사용하지 않는 best-soeat-card 컴포넌트 및 관련 파일 삭제
 - SosoTalkCardCompact를 best-sosotalk-section으로 이동
 - 스크롤 불필요 시 wheel 이벤트 preventdefault 스킵 처리
