--- src/widgets/home/ui/best-soeat-card/best-soeat-card.types.ts
@@ -1,8 +1,0 @@
-export interface BestSoeatCardProps {
-  id: number;
-  title: string;
-  region: string;
-  meetingAt: string;
-  thumbnailUrl?: string;
-  thumbnailAlt?: string;
-}
