--- src/widgets/home/ui/best-soeat-card/index.ts
@@ -1,3 +1,0 @@
-export { BestSoeatCard } from './best-soeat-card';
-export type { BestSoeatCardProps } from './best-soeat-card.types';
-export { SosoTalkCardCompact } from './sosotalk-card-compact';
