+++ src/widgets/home/ui/best-sosotalk-section/sosotalk-card-compact.tsx
